fix+test(server/core/client/vnet): round-10 review — 2 HIGH + 8 MEDIUM + 6 LOW, all fixed
Production fixes:
- HIGH: yamux-stream control reads bounded by CTL_READ_TIMEOUT 30s
  (post-auth half-frame trickle pinned task + fd + semaphore permit +
  run_id registration; Go bounds via heartbeatWorker ctl.Close())
- HIGH: vnet route advertise membership guard (advertiser must hold a
  vnet proxy in the target virtual_net) + 64-route per-client cap
  (new keys refused at cap, updates to owned keys allowed)
- MEDIUM: XTCP raw-KCP path (no tcp-mux) HWM-parked writer now wakes
  on notify OR 10ms timer and forces a drive_kcp drain — no permanent
  write stall with a silent peer; poll_flush notifies when it drained
  a full buffer
- MEDIUM: vhost dead code removed (location_routes/by_proxy_locations/
  lookup_by_path) — Go routes locations only domain-scoped;
  lookup_combined = lookup_wildcard (-159 lines)
- MEDIUM: h2c vhost + client h2 plugin chunk body streamed in 64KiB
  slices (attacker-declared huge chunk no longer allocs read_exact(size))
- MEDIUM: vnet controller distinguishes full vs failed control queue
  via ControlSink::is_failed() — transient full no longer kills TUN pump
- MEDIUM: tun_macos AF_INET6 = 30 (Linux is 10), family from first-byte
  version nibble
- MEDIUM: frpc health defaults Go parity (health.go:57-64 sets defaults
  only when <= 0, not a .max(10) floor); client server_addr default
  "0.0.0.0" (client.go:86); visitor validation mirrors Go
  validation/visitor.go:42-63 (name/serverName required, bindPort==0
  rejected, negative = no-bind passes, XTCP protocol kcp|quic)
- MEDIUM: StartWorkConn src_port/dst_port Option<i32> → Option<u16>
  (Go uint16 — hostile -1/70000 frames now fail deserialization)
- LOW: quic max_idle_timeout VarInt overflow leaves quinn default
  instead of panicking (panic=abort); OIDC jti 4KiB cap; ReplayTable
  expiry saturating_add; login rejects run_id Some("")

Tests (+4: 1367 → 1371):
- msg::test_start_work_conn_ports_are_u16 (hostile frame rejection)
- config::test_visitor_config_validation_matches_go (4 reject + no-bind pass)
- nathole::advertise_without_vnet_proxy_is_refused (membership guard)
- nathole::advertise_respects_per_client_route_cap (65th refused,
  owned-key update allowed); existing vnet tests gain register_vnet_proxy
  fixture so the guard does not short-circuit their assertions

Verified: cargo check --workspace + --all-features clean; fmt clean;
clippy -D warnings clean; 1371 passed / 0 failed.

+++ frp-client/src/work_conn.rs
@@ -1680,8 +1680,8 @@ pub(crate) fn spawn_work_conn(cfg: WorkConnConfig) -> tokio::task::JoinHandle<()
                                             frp_core::proxy_protocol::build_proxy_protocol_v1(
                                                 src,
                                                 swc.dst_addr.as_deref().unwrap_or("0.0.0.0"),
-                                                swc.src_port.unwrap_or(0) as u16,
-                                                swc.dst_port.unwrap_or(0) as u16,
+                                                swc.src_port.unwrap_or(0),
+                                                swc.dst_port.unwrap_or(0),
                                             );
                                         if let Err(e) = local.write_all(header.as_bytes()).await {
                                             warn!(error = %e, "Failed to write PROXY v1 header: {}", e);
@@ -1690,8 +1690,8 @@ pub(crate) fn spawn_work_conn(cfg: WorkConnConfig) -> tokio::task::JoinHandle<()
                                         match frp_core::proxy_protocol::build_proxy_protocol_v2(
                                             src,
                                             swc.dst_addr.as_deref().unwrap_or("0.0.0.0"),
-                                            swc.src_port.unwrap_or(0) as u16,
-                                            swc.dst_port.unwrap_or(0) as u16,
+                                            swc.src_port.unwrap_or(0),
+                                            swc.dst_port.unwrap_or(0),
                                         ) {
                                             Ok(header) => {
                                                 if let Err(e) = local.write_all(&header).await {
